Split Bindgen into separate components
